fixed git issues? fixed commands/slash/nice issue
